[*] Refactor include directory, create internal header directory and update style

+++ src/_x86/irq.c
@@ -1,5 +1,8 @@
 #include <smkos/kernel.h>
-#include <smkos/core.h>
+#include <smkos/arch.h>
+#include <smkos/kstruct/map.h>
+#include <smkos/kstruct/task.h>
+// #include <smkos/core.h>
 
 /* ----------------------------------------------------------------------- */
 /** Hardware Exception:
@@ -52,9 +55,9 @@ void(*x86_irq_hanlder[16])();
     IRQ 14 : Primary HDD
     IRQ 15 : Secondary HDD
 */
-void sys_irq (int no, size_t* params)
+void sys_irq (int no, size_t *params)
 {
-  kCPU.current_->stackPtr_ = params;
+  kCPU.current_->stackPtr_ = (size_t)params;
 
   if (no < 0 || no >= 16)
     kpanic ("IRQ no %d !?\n", no);
@@ -72,9 +75,9 @@ void sys_irq (int no, size_t* params)
 int system_call (int no, size_t p1, size_t p2, size_t p3, size_t p4, size_t p5);
 
 
-void sys_call(size_t* params)
+void sys_call(size_t *params)
 {
-  kCPU.current_->stackPtr_ = params;
+  kCPU.current_->stackPtr_ = (size_t)params;
   int err = system_call(params[11], params[10], params[9], params[8], params[7], params[6]);
   params[11] = err;
   params[9] = __geterrno();
